Add immediate data representation

+++ lib/fabrics/ofi/src/internal/Endpoint.hpp
@@ -155,7 +155,7 @@ namespace mxl::lib::fabrics::ofi
          * \param 64 bits of user data that will be available in the completion entry associated with this transfer.
          */
         void write(LocalRegionGroup const& localGroup, RemoteRegionGroup const& remoteGroup, ::fi_addr_t destAddr = FI_ADDR_UNSPEC,
-            std::optional<std::uint64_t> immData = std::nullopt);
+            std::optional<std::uint32_t> immData = std::nullopt);
 
         /*
          * Push a recv work request to the endpoint work queue. In the MXL context the memory region passed here is not the memory
